+++ src/Model.hs
@@ -16,6 +16,7 @@ import CustomTypes
 import Vehicles.Components
 import People.Data
 import Research.Data
+import Resources ( RawResource(..), Biological(..), Chemical(..), Mechanical(..) )
 
 -- You can define all of your database entities in the entities file.
 -- You can find more information on persistent and how to declare entities